renamed customSelectors folder to "Selectors" and moved it into store folder
