+++ assignment2/src/store/selectors/StocksSelectors.js
